requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
boto3==1.17.49
botocore==1.20.49
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-bootstrap4==3.0.1
django-phonenumber-field==5.0.0
django-storages==1.11.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
djangorestframework-stubs==1.4.0
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
mypy==0.812
mypy-extensions==0.4.3
mysqlclient==2.0.3
phonenumbers==8.12.21
Pillow==8.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
twilio==6.56.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
